feat(damka): refactor game logic to use Zustand store for state management
